fix: restore ZoneEngine hysteresis, batch DAO, and misc improvements
Recovered from APK decompilation.

- ZoneEngine: add 2 BPM hysteresis to prevent zone oscillation when
  HR hovers near a threshold. Tracks lastStatus to widen thresholds
  when already IN_ZONE.
- TrackPointDao: add getTrackPointsForWorkouts batch query
- BootcampUiState: add maxHr field, populated from UserProfileRepository
- Formatters: add String.asModeLabel() extension for display formatting
- FitnessLoadCalculator: compute CTL/ATL decay in double precision
  before casting to float for improved accuracy

+++ app/src/main/java/com/hrcoach/domain/engine/FitnessLoadCalculator.kt
@@ -10,8 +10,8 @@ object FitnessLoadCalculator {
         val tsb: Float
     )
 
-    private const val CTL_TAU = 42f
-    private const val ATL_TAU = 7f
+    private const val CTL_TAU = 42.0
+    private const val ATL_TAU = 7.0
 
     fun updateLoads(
         currentCtl: Float,
@@ -19,12 +19,12 @@ object FitnessLoadCalculator {
         trimpScore: Float,
         daysSinceLast: Int
     ): LoadResult {
-        val days = daysSinceLast.coerceAtLeast(0).toFloat().coerceAtLeast(0.5f)
-        val ctlDecay = exp(-days / CTL_TAU).toFloat()
-        val atlDecay = exp(-days / ATL_TAU).toFloat()
+        val days = daysSinceLast.coerceAtLeast(0).toDouble().coerceAtLeast(0.5)
+        val ctlDecay = exp(-days / CTL_TAU)
+        val atlDecay = exp(-days / ATL_TAU)
 
-        val newCtl = currentCtl * ctlDecay + trimpScore * (1f - ctlDecay)
-        val newAtl = currentAtl * atlDecay + trimpScore * (1f - atlDecay)
+        val newCtl = (currentCtl * ctlDecay + trimpScore * (1.0 - ctlDecay)).toFloat()
+        val newAtl = (currentAtl * atlDecay + trimpScore * (1.0 - atlDecay)).toFloat()
 
         return LoadResult(
             ctl = newCtl,
